+++ packages/openactive-broker-microservice/tsconfig.json
@@ -10,6 +10,8 @@
   },
   "include": [
     "app.js",
+    "src/listeners/**/*.js",
+    "src/models/**/*.d.ts",
     "src/util/**/*.js",
     "src/*.js"
   ]
